+++ Back-end/src/main/java/com/chatop/dtos/MessageRequest.java
@@ -7,7 +7,7 @@ import lombok.Setter;
 @Setter
 public class MessageRequest {
     private String message;
-    private Long rental_id;
-    private Long user_id;
+    private Integer rental_id;
+    private Integer user_id;
 }
